// backend/src/components/product/interfaces/Inventory.ts
// Interface automatically generated by schemas-to-ts

import { ProductColor } from '../../../api/product-color/content-types/product-color/product-color';
import { ProductSize } from '../../../api/product-size/content-types/product-size/product-size';
import { ProductColor_Plain } from '../../../api/product-color/content-types/product-color/product-color';
import { ProductSize_Plain } from '../../../api/product-size/content-types/product-size/product-size';

export interface Inventory {
  product_color?: { data: ProductColor };
  product_size?: { data: ProductSize };
  quantity: number;
}
export interface Inventory_Plain {
  product_color?: ProductColor_Plain;
  product_size?: ProductSize_Plain;
  quantity: number;
}

export interface Inventory_NoRelations {
  product_color?: number;
  product_size?: number;
  quantity: number;
}